+++ src/state/pools/index.ts
@@ -44,7 +44,7 @@ export const fetchPoolsPublicDataAsync = (currentBlock: number) => async (dispat
 
   const liveData = poolsConfig.map((pool) => {
     const blockLimit = blockLimits.find((entry) => entry.sousId === pool.sousId)
-    console.log('limit', blockLimits)
+    // console.log('limit', blockLimits)
     const totalStaking = totalStakings.find((entry) => entry.sousId === pool.sousId)
     const isPoolEndBlockExceeded = currentBlock > 0 && blockLimit ? currentBlock > Number(blockLimit.endBlock) : false
     const isPoolFinished = pool.isFinished || isPoolEndBlockExceeded
